Create Android/Apple SDK updates as draft (#635)
As a safety feature, GitHub does not run workflows on PRs created by other workflows, so the workaround is to use a draft and once a human marks draft as ready then the tests will run.

Also, used `github.repository_owner` instead of hardcoding embrace-io so it is safer for this to run in forks.